Fix icon state checker script broken by mirror bot
