+++ test/check_icon_state_limit.py
@@ -1,24 +1,33 @@
 import argparse, re, sys
 from os import path, walk
-import dmitool # This import is why this script is here. If someone can import this file cleanly from [repo root]/test/ instead, feel free
+from PIL import Image
 
 opt = argparse.ArgumentParser()
 opt.add_argument('dir', help='The directory to scan for *.dmi files with an excess number of icon states.')
 args = opt.parse_args()
 
+STATE_PATTERN = r'^state\s*='
+
+def get_states_count(path):
+    try:
+        im = Image.open(path)
+        return len(re.findall(STATE_PATTERN, im.info["Description"], re.MULTILINE))
+    except (Image.UnidentifiedImageError):
+        print("{0} is not a valid image".format(path))
+        exit(1)
+
 if(not path.isdir(args.dir)):
     print('Not a directory')
     sys.exit(1)
-  
+
 bad_dmi_files = []
-  
+
 # This section parses all *.dmi files in the given directory, recursively.
 for root, subdirs, files in walk(args.dir):
     for filename in files:
         if filename.endswith('.dmi'):
             file_path = path.join(root, filename)
-            dmi_info = dmitool.info(file_path)
-            number_of_icon_states = len(dmi_info["states"])
+            number_of_icon_states = get_states_count(file_path)
             if number_of_icon_states > 512:
                 bad_dmi_files.append((file_path, number_of_icon_states))
